@@ -2,13 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Voronkovich\SberbankAcquiring\Exception;
+namespace Gruzoveek\SberbankAcquiring\Exception;
 
-/**
- * Network exception.
- *
- * @author Oleg Voronkovich <oleg-voronkovich@yandex.ru>
- */
+
 class NetworkException extends SberbankAcquiringException
 {
 }
