@@ -2,11 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Voronkovich\SberbankAcquiring\Exception;
+namespace Gruzoveek\SberbankAcquiring\Exception;
 
-/**
- * @author Oleg Voronkovich <oleg-voronkovich@yandex.ru>
- */
+
 class ResponseParsingException extends SberbankAcquiringException
 {
 }
